partial install added Link checker added

+++ install.sh
@@ -1,11 +1,22 @@
 #!/usr/bin/env bash
 
+ARGS=$#
+ARG=$1
+ONLY=$2
+
+if [[ "$ARG" == "--help" ]]; then
+    echo "Usage: bash install.sh [--only \"test1 test2\"]"
+    exit
+fi
+
 install() {
-    rm -rf $1
-    git clone https://github.com/SecuritySquad/$1.git
-    cd $1
-    sh install.sh
-    cd ..
+    if [[ ${ARGS} -eq 0 || "$ARG" == "--only" && "$ONLY" =~ "$1" ]]; then
+        rm -rf $1
+        git clone https://github.com/SecuritySquad/$1.git
+        cd $1
+        bash install.sh
+        cd ..
+    fi
 }
 
 cd ..
@@ -17,6 +28,8 @@ install webifier-test-virusscan
 install webifier-test-header-inspection
 
 install webifier-test-portscan
+
+install webifier-test-linkchecker
 
 docker rmi $(docker images --filter "dangling=true" -q)
